Rename `services/bitcoin/context`
to `services/bitcoin/index` to have similar structure for all services
based on https://github.com/thorchain/asgardex-electron/pull/614
