+++ src/renderer/services/chain/fees/stake.ts
@@ -12,7 +12,7 @@ import { liveData } from '../../../helpers/rx/liveData'
 import { observableState } from '../../../helpers/stateHelper'
 import { StakeType } from '../../../types/asgardex'
 import * as BNB from '../../binance/service'
-import * as BTC from '../../bitcoin/context'
+import * as BTC from '../../bitcoin'
 import { selectedPoolAsset$, selectedPoolChain$ } from '../../midgard/common'
 import { symDepositAssetTxMemo$, asymDepositTxMemo$ } from '../memo'
 import { FeeLD, LoadFeesHandler, StakeFeesLD } from '../types'
