+++ src/renderer/services/wallet/balances.ts
@@ -15,7 +15,7 @@ import { sequenceTOptionFromArray } from '../../helpers/fpHelpers'
 import { liveData } from '../../helpers/rx/liveData'
 import { network$ } from '../app/service'
 import * as BNB from '../binance/service'
-import * as BTC from '../bitcoin/context'
+import * as BTC from '../bitcoin'
 import * as ETH from '../ethereum/service'
 import { selectedAsset$ } from './common'
 import { INITIAL_ASSETS_WB_STATE } from './const'
